refactor: architecture realignment - monitoring/config migration to module-infra (#385, #386)
## Phase 0: DIP Compliance
- Create BufferStatusQuery Port interface in module-core
- RedisBufferRepositoryImpl now implements BufferStatusQuery
- RedisMetricsCollector uses BufferStatusQuery instead of RedisBufferRepository
- MonitoringAlertService uses BufferStatusQuery instead of RedisBufferRepository

## Phase 1: Remove Duplicate Monitoring
- Delete module-app/src/main/kotlin/maple/expectation/monitoring/ (9 Kotlin files)
- Update imports to use module-infra monitoring classes

## Phase 2: Remove Duplicate Config
- Delete 11 Java config files from module-app/config:
  - BufferConfig.java, BufferProperties.java
  - BatchConfig.java, BatchProperties.java
  - LikeSyncConfig.java, LikeBufferConfig.java, LikeRealtimeSyncConfig.java
  - EventConsumerConfig.java, OpenTelemetryConfig.java
  - MaplestoryApiConfig.java, EquipmentProcessingExecutorConfig.java
- Update all imports to use module-infra config classes
- Fix Kotlin property access to use Java getter methods

## Verification
- All tests pass
- No reverse dependency (module-infra does not depend on module-app)
- DIP compliance verified (monitoring uses Port interface)
