--- module-app/src/main/kotlin/maple/expectation/monitoring/MonitoringAlertService.kt
@@ -1,73 +1,0 @@
-package maple.expectation.monitoring
-
-import maple.expectation.domain.repository.RedisBufferRepository
-import maple.expectation.error.CommonErrorCode
-import maple.expectation.error.exception.MonitoringException
-import maple.expectation.infrastructure.alert.StatelessAlertService
-import maple.expectation.infrastructure.executor.LogicExecutor
-import maple.expectation.infrastructure.executor.TaskContext
-import maple.expectation.infrastructure.lock.LockStrategy
-import maple.expectation.infrastructure.config.MonitoringThresholdProperties
-import org.slf4j.LoggerFactory
-import org.springframework.scheduling.annotation.Scheduled
-import org.springframework.stereotype.Component
-
-@Component
-class MonitoringAlertService(
-    private val redisBufferRepository: RedisBufferRepository,
-    private val statelessAlertService: StatelessAlertService,
-    private val lockStrategy: LockStrategy,
-    private val executor: LogicExecutor,
-    private val thresholdProperties: MonitoringThresholdProperties
-) {
-
-  private val log = LoggerFactory.getLogger(MonitoringAlertService::class.java)
-
-  /**
-   * 버퍼 포화도 체크 로직 (Leader Election 패턴)
-   *
-   * tryLockImmediately()를 사용하여 예외 없이 리더 선출. 락 획득 실패(Follower)는 정상 시나리오이므로 조용히 스킵.
-   */
-  @Scheduled(fixedRate = 5000)
-  fun checkBufferSaturation() {
-    val context = TaskContext.of("Monitoring", "CheckSaturation")
-
-    val isLeader = lockStrategy.tryLockImmediately("global-monitoring-lock", 4)
-
-    if (!isLeader) {
-      log.debug("⏭️ [Monitoring] 리더 선출 실패 - 다른 인스턴스가 리더입니다. 체크 스킵.")
-      return
-    }
-
-    executor.executeOrCatch(
-        { performBufferCheck(); null },
-        { t -> handleMonitoringFailure(t) },
-        context
-    )
-  }
-
-  /** 헬퍼 1: 실제 수치 확인 및 알림 로직 (로직 응집도 향상) */
-  private fun performBufferCheck() {
-    val globalPending = redisBufferRepository.getTotalPendingCount()
-
-    if (globalPending > thresholdProperties.bufferSaturationCount) {
-      val exception = MonitoringException(CommonErrorCode.SYSTEM_CAPACITY_EXCEEDED, globalPending)
-
-      executor.executeVoidJava(
-          {
-            statelessAlertService.sendCritical(
-                "🚨 GLOBAL BUFFER SATURATION", exception.message ?: "Unknown error", exception
-            )
-            log.warn("[{}] {}", exception.errorCode.code, exception.message)
-          },
-          TaskContext.of("Alert", "SendDiscord", globalPending.toString())
-      )
-    }
-  }
-
-  /** 모니터링 실패 대응 (실제 장애만 로깅) */
-  private fun handleMonitoringFailure(t: Throwable): Void? {
-    log.error("❌ [Monitoring] 버퍼 모니터링 중 장애 발생: {}", t.message, t)
-    return null
-  }
-}
