+++ module-app/src/main/kotlin/maple/expectation/service/ingestion/BatchWriter.kt
@@ -2,7 +2,7 @@ package maple.expectation.service.ingestion
 
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
-import maple.expectation.config.BatchProperties
+import maple.expectation.infrastructure.config.BatchProperties
 import maple.expectation.core.port.out.MessageQueue
 import maple.expectation.domain.event.IntegrationEvent
 import maple.expectation.domain.nexon.NexonApiCharacterData
