+++ module-app/src/test/java/maple/expectation/monitoring/MetricsCollectorTest.java
@@ -8,10 +8,10 @@ import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import java.util.Map;
-import maple.expectation.monitoring.collector.CircuitBreakerMetricsCollector;
-import maple.expectation.monitoring.collector.GoldenSignalsCollector;
-import maple.expectation.monitoring.collector.JvmMetricsCollector;
-import maple.expectation.monitoring.collector.MetricCategory;
+import maple.expectation.infrastructure.monitoring.collector.CircuitBreakerMetricsCollector;
+import maple.expectation.infrastructure.monitoring.collector.GoldenSignalsCollector;
+import maple.expectation.infrastructure.monitoring.collector.JvmMetricsCollector;
+import maple.expectation.infrastructure.monitoring.collector.MetricCategory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
